Move existing tests and update them
